refactor: move Search and POI content to dedicated folder
Plus, already move other content to temporary src/content/index
