+++ src/content/generics.tsx
@@ -1,0 +1,7 @@
+export const SHADE_SUPPORT_NOTE = (
+  <p className="text-xs">
+    Leider können die Schatten auf diesem Endgerät oder in diesem Browser nicht
+    dargestellt werden. Bitte versuche ein anderes Endgerät oder einen anderen
+    Browser, um die Karte zu öffnen.
+  </p>
+)
